Move TypePatternExpr type to PatternExpr
